Move test outside abstract parent
